reducers refactored, chat and file actions
