+++ client/src/StreamComponents/FileDirectory/FileDirectory.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Tree } from 'antd';
 import { connect } from 'react-redux';
-import { updateFile } from '../../redux/ducks/directoryDuck';
+import { fileActions } from '../../redux/_actions';
+// import { updateFile } from '../../redux/ducks/directoryDuck';
 
 const TreeNode = Tree.TreeNode;
 
@@ -70,7 +71,7 @@ const mapStateToProps = (state) => {
 
 const mapDispatchToProps = (dispatch) => {
   return {
-    sendFileID: (fileID) => dispatch(updateFile(fileID))
+    sendFileID: (fileID) => dispatch(fileActions.updateFile(fileID))
   }
 }
