+++ client/src/redux/_actions/file.actions.js
@@ -1,0 +1,7 @@
+import { fileConstants } from '../_constants';
+
+export const fileActions = {
+  updateFile
+}
+
+const updateFile = (fileID) => ({ type: fileConstants.FILE_REQUEST, payload: { fileID } });
