+++ client/src/redux/_actions/index.js
@@ -1,0 +1,2 @@
+export * from './chat.actions.js';
+export * from './file.actions.js';
