--- client/src/redux/_constants/directory.constants.js
@@ -1,4 +1,0 @@
-export const directoryConstants = {
-  FILE_REQUEST:     'server/file_change',
-  DIRECTORY_UPDATE: 'DIRECTORY_UPDATE'
-}
