+++ client/src/redux/_constants/file.constants.js
@@ -1,3 +1,7 @@
 export const fileConstants = {
-  FILE_UPDATE: 'FILE_UPDATE'
+  FILE_UPDATE: 'FILE_UPDATE',
+  DIRECTORY_UPDATE: 'DIRECTORY_UPDATE',
+
+  FILE_REQUEST:     'server/file_change'
+  
 }
