+++ client/src/redux/_constants/index.js
@@ -1,4 +1,3 @@
 export * from './chat.constants.js';
-export * from './directory.constants.js';
 export * from './file.constants.js';
 export * from './stream.constants.js';
