+++ client/src/redux/ducks/Ducktator.js
@@ -1,14 +1,14 @@
-import { combineReducers } from 'redux'
+// import { combineReducers } from 'redux'
 
-import { chatReducer }  from './chatDuck'
-import { directoryReducer } from './directoryDuck'
-import { fileReducer } from './fileDuck'
-import { streamsReducer } from './streamsDuck'
+// import { chatReducer }  from './chatDuck'
+// import { directoryReducer } from './directoryDuck'
+// import { fileReducer } from './fileDuck'
+// import { streamsReducer } from './streamsDuck'
 
 
-export const rootReducer = combineReducers({
-  chat: chatReducer,
-  directory: directoryReducer,
-  file: fileReducer,
-  streams: streamsReducer
-})
+// export const rootReducer = combineReducers({
+//   chat: chatReducer,
+//   directory: directoryReducer,
+//   file: fileReducer,
+//   streams: streamsReducer
+// })
