+++ client/src/redux/ducks/directoryDuck.js
@@ -1,4 +1,4 @@
-import { directoryConstants } from '../_constants';
+// import { fileConstants } from '../_constants';
 
 // delete delete delete
 // Outgoing
@@ -7,18 +7,18 @@ import { directoryConstants } from '../_constants';
 // const DIRECTORY_UPDATE = 'DIRECTORY_UPDATE';
 
 // Action Creators
-export const updateFile = (fileID) => ({ type: directoryConstants.FILE_REQUEST, payload: { fileID } });
+// export const updateFile = (fileID) => ({ type: fileConstants.FILE_REQUEST, payload: { fileID } });
 
-// DIRECTORY REDUCER
-const directoryState = { directoryStructure: { 'staticKey': {fileNames:['Your Great App'],fileIDs:['uuid']}  }};
+// // DIRECTORY REDUCER
+// const directoryState = { directoryStructure: { 'staticKey': {fileNames:['Your Great App'],fileIDs:['uuid']}  }};
 
-export const directoryReducer = (state = directoryState, action) => {
-  switch (action.type) {
-    case directoryConstants.DIRECTORY_UPDATE:
-      console.log('directory_update action', action.payload)
-      return { ...state, directoryStructure: action.payload };
+// export const directoryReducer = (state = directoryState, action) => {
+//   switch (action.type) {
+//     case directoryConstants.DIRECTORY_UPDATE:
+//       console.log('directory_update action', action.payload)
+//       return { ...state, directoryStructure: action.payload };
     
-    default:
-      return state;
-  }
-}
+//     default:
+//       return state;
+//   }
+// }
