+++ client/src/redux/ducks/streamsDuck.js
@@ -15,26 +15,26 @@ export const updateActiveStreams = (activeStreams) => ({ type: streamConstants.A
 
 export const updateArchivedStreams = (archivedStreams) => ({ type: streamConstants.ARCHIVED_STREAMS_UPDATE, payload: archivedStreams  });
 
-// STREAM REDUCER
+// // STREAM REDUCER
 
-export const streamsReducer = (state = { activeStreams: ['noStreams'], scheduledStreams: ['noStreams'], archivedStreams: ['noStreams'] }, action) => {
-  switch(action.type) {
-    case streamConstants.SCHEDULED_STREAMS_UPDATE:
-      console.log('userstreams update', action.payload);
-      return { ...state, scheduledStreams: action.payload }
+// export const streamsReducer = (state = { activeStreams: ['noStreams'], scheduledStreams: ['noStreams'], archivedStreams: ['noStreams'] }, action) => {
+//   switch(action.type) {
+//     case streamConstants.SCHEDULED_STREAMS_UPDATE:
+//       console.log('userstreams update', action.payload);
+//       return { ...state, scheduledStreams: action.payload }
 
-    case streamConstants.ACTIVE_STREAMS_UPDATE:
-    console.log('userstreams update', action.payload);
-    return { ...state, activeStreams: action.payload }
+//     case streamConstants.ACTIVE_STREAMS_UPDATE:
+//     console.log('userstreams update', action.payload);
+//     return { ...state, activeStreams: action.payload }
 
-    case streamConstants.ARCHIVED_STREAMS_UPDATE:
-    console.log('userstreams update', action.payload);
-    return { ...state, archivedStreams: action.payload }
+//     case streamConstants.ARCHIVED_STREAMS_UPDATE:
+//     console.log('userstreams update', action.payload);
+//     return { ...state, archivedStreams: action.payload }
 
-    default:
-      return state;
-  }
-};
+//     default:
+//       return state;
+//   }
+// };
 
 // AJAX ACTIONS
 export const fetchScheduledStreams = () => {
